requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
Jinja2==3.1.6
MarkupSafe==3.0.2
sqlparse==0.5.5
